array type changed

+++ src/dct.cc
@@ -2,12 +2,12 @@
 
 namespace vp8 {
 
-void DCT(std::array<std::array<int, 4>, 4> &subblock) {
+void DCT(std::array<std::array<int16_t, 4>, 4> &subblock) {
   for (int i = 0; i < 4; i++) {
-    int a = (subblock[i][0] + subblock[i][3]) << 3;
-    int b = (subblock[i][1] + subblock[i][2]) << 3;
-    int c = (subblock[i][1] - subblock[i][2]) << 3;
-    int d = (subblock[i][0] - subblock[i][3]) << 3;
+    int a = ((int)subblock[i][0] + (int)subblock[i][3]) << 3;
+    int b = ((int)subblock[i][1] + (int)subblock[i][2]) << 3;
+    int c = ((int)subblock[i][1] - (int)subblock[i][2]) << 3;
+    int d = ((int)subblock[i][0] - (int)subblock[i][3]) << 3;
 
     subblock[i][0] = a + b;
     subblock[i][2] = a - b;
@@ -16,10 +16,10 @@ void DCT(std::array<std::array<int, 4>, 4> &subblock) {
   }
 
   for (int i = 0; i < 4; i++) {
-    int a = subblock[0][i] + subblock[3][i];
-    int b = subblock[1][i] + subblock[2][i];
-    int c = subblock[1][i] - subblock[2][i];
-    int d = subblock[0][i] - subblock[3][i];
+    int a = (int)subblock[0][i] + (int)subblock[3][i];
+    int b = (int)subblock[1][i] + (int)subblock[2][i];
+    int c = (int)subblock[1][i] - (int)subblock[2][i];
+    int d = (int)subblock[0][i] - (int)subblock[3][i];
 
     subblock[0][i] = (a + b + 7) >> 4;
     subblock[2][i] = (a - b + 7) >> 4;
@@ -27,12 +27,12 @@ void DCT(std::array<std::array<int, 4>, 4> &subblock) {
     subblock[3][i] = (d * 2217 - c * 5352 + 51000) >> 16;
   }
 }
-void WHT(std::array<std::array<int, 4>, 4> &subblock) {
+void WHT(std::array<std::array<int16_t, 4>, 4> &subblock) {
   for (int i = 0; i < 4; i++) {
-    int a = (subblock[i][0] + subblock[i][2]) << 2;
-    int d = (subblock[i][1] + subblock[i][3]) << 2;
-    int c = (subblock[i][1] - subblock[i][3]) << 2;
-    int b = (subblock[i][0] - subblock[i][2]) << 2;
+    int a = ((int)subblock[i][0] + (int)subblock[i][2]) << 2;
+    int d = ((int)subblock[i][1] + (int)subblock[i][3]) << 2;
+    int c = ((int)subblock[i][1] - (int)subblock[i][3]) << 2;
+    int b = ((int)subblock[i][0] - (int)subblock[i][2]) << 2;
     subblock[i][0] = a + d + (a != 0 ? 1 : 0);
     subblock[i][1] = b + c;
     subblock[i][2] = b - c;
@@ -40,10 +40,10 @@ void WHT(std::array<std::array<int, 4>, 4> &subblock) {
   }
 
   for (int i = 0; i < 4; i++) {
-    int a = subblock[0][i] + subblock[2][i];
-    int d = subblock[1][i] + subblock[3][i];
-    int c = subblock[1][i] - subblock[3][i];
-    int b = subblock[0][i] - subblock[2][i];
+    int a = (int)subblock[0][i] + (int)subblock[2][i];
+    int d = (int)subblock[1][i] + (int)subblock[3][i];
+    int c = (int)subblock[1][i] - (int)subblock[3][i];
+    int b = (int)subblock[0][i] - (int)subblock[2][i];
 
     int a2 = a + d;
     int b2 = b + c;
@@ -65,16 +65,16 @@ void WHT(std::array<std::array<int, 4>, 4> &subblock) {
 static const int cospi8_sqrt2_minus1 = 20091;
 static const int sinpi8_sqrt2 = 35468;
 
-void IDCT(std::array<std::array<int, 4>, 4> &subblock) {
+void IDCT(std::array<std::array<int16_t, 4>, 4> &subblock) {
   for (int i = 0; i < 4; i++) {
-    int a = subblock[0][i] + subblock[2][i];
-    int b = subblock[0][i] - subblock[2][i];
+    int a = (int)subblock[0][i] + (int)subblock[2][i];
+    int b = (int)subblock[0][i] - (int)subblock[2][i];
 
-    int tmp1 = (subblock[1][i] * sinpi8_sqrt2) >> 16;
-    int tmp2 = subblock[3][i] + ((subblock[3][i] * cospi8_sqrt2_minus1) >> 16);
+    int tmp1 = ((int)subblock[1][i] * sinpi8_sqrt2) >> 16;
+    int tmp2 = (int)subblock[3][i] + (((int)subblock[3][i] * cospi8_sqrt2_minus1) >> 16);
     int c = tmp1 - tmp2;
-    tmp1 = subblock[1][i] + ((subblock[1][i] * cospi8_sqrt2_minus1) >> 16);
-    tmp2 = (subblock[3][i] * sinpi8_sqrt2) >> 16;
+    tmp1 = (int)subblock[1][i] + (((int)subblock[1][i] * cospi8_sqrt2_minus1) >> 16);
+    tmp2 = ((int)subblock[3][i] * sinpi8_sqrt2) >> 16;
     int d = tmp1 + tmp2;
 
     subblock[0][i] = a + d;
@@ -84,14 +84,14 @@ void IDCT(std::array<std::array<int, 4>, 4> &subblock) {
   }
 
   for (int i = 0; i < 4; i++) {
-    int a = subblock[i][0] + subblock[i][2];
-    int b = subblock[i][0] - subblock[i][2];
+    int a = (int)subblock[i][0] + (int)subblock[i][2];
+    int b = (int)subblock[i][0] - (int)subblock[i][2];
 
-    int tmp1 = (subblock[i][1] * sinpi8_sqrt2) >> 16;
-    int tmp2 = subblock[i][3] + ((subblock[i][3] * cospi8_sqrt2_minus1) >> 16);
+    int tmp1 = ((int)subblock[i][1] * sinpi8_sqrt2) >> 16;
+    int tmp2 = (int)subblock[i][3] + (((int)subblock[i][3] * cospi8_sqrt2_minus1) >> 16);
     int c = tmp1 - tmp2;
-    tmp1 = subblock[i][1] + ((subblock[i][1] * cospi8_sqrt2_minus1) >> 16);
-    tmp2 = (subblock[i][3] * sinpi8_sqrt2) >> 16;
+    tmp1 = (int)subblock[i][1] + (((int)subblock[i][1] * cospi8_sqrt2_minus1) >> 16);
+    tmp2 = ((int)subblock[i][3] * sinpi8_sqrt2) >> 16;
     int d = tmp1 + tmp2;
 
     subblock[i][0] = (a + d + 4) >> 3;
@@ -101,12 +101,12 @@ void IDCT(std::array<std::array<int, 4>, 4> &subblock) {
   }
 }
 
-void IWHT(std::array<std::array<int, 4>, 4> &subblock) {
+void IWHT(std::array<std::array<int16_t, 4>, 4> &subblock) {
   for (int i = 0; i < 4; i++) {
-    int a = subblock[0][i] + subblock[3][i];
-    int b = subblock[1][i] + subblock[2][i];
-    int c = subblock[1][i] - subblock[2][i];
-    int d = subblock[0][i] - subblock[3][i];
+    int a = (int)subblock[0][i] + (int)subblock[3][i];
+    int b = (int)subblock[1][i] + (int)subblock[2][i];
+    int c = (int)subblock[1][i] - (int)subblock[2][i];
+    int d = (int)subblock[0][i] - (int)subblock[3][i];
 
     subblock[0][i] = a + b;
     subblock[1][i] = c + d;
@@ -115,10 +115,10 @@ void IWHT(std::array<std::array<int, 4>, 4> &subblock) {
   }
 
   for (int i = 0; i < 4; i++) {
-    int a = subblock[i][0] + subblock[i][3];
-    int b = subblock[i][1] + subblock[i][2];
-    int c = subblock[i][1] - subblock[i][2];
-    int d = subblock[i][0] - subblock[i][3];
+    int a = (int)subblock[i][0] + (int)subblock[i][3];
+    int b = (int)subblock[i][1] + (int)subblock[i][2];
+    int c = (int)subblock[i][1] - (int)subblock[i][2];
+    int d = (int)subblock[i][0] - (int)subblock[i][3];
 
     subblock[i][0] = (a + b + 3) >> 3;
     subblock[i][1] = (c + d + 3) >> 3;
